test(GHA): extract setup/teardowns into common lib
Feedback: https://github.com/readmeio/rdme/pull/577#discussion_r961377510
