+++ __tests__/cmds/openapi/index.test.ts
@@ -1,6 +1,4 @@
 /* eslint-disable no-console */
-import type { Response } from 'simple-git';
-
 import fs from 'fs';
 
 import chalk from 'chalk';
@@ -11,11 +9,8 @@ import prompts from 'prompts';
 import OpenAPICommand from '../../../src/cmds/openapi';
 import SwaggerCommand from '../../../src/cmds/swagger';
 import APIError from '../../../src/lib/apiError';
-import configstore from '../../../src/lib/configstore';
-import { git } from '../../../src/lib/createGHA';
-import * as createGHAObject from '../../../src/lib/createGHA';
 import getAPIMock, { getAPIMockWithVersionHeader } from '../../helpers/get-api-mock';
-import getGitRemoteMock from '../../helpers/get-git-mock';
+import { after, before } from '../../helpers/get-gha-setup';
 
 const openapi = new OpenAPICommand();
 const swagger = new SwaggerCommand();
@@ -995,36 +990,14 @@ describe('rdme openapi', () => {
     let yamlOutput;
 
     beforeEach(() => {
-      consoleInfoSpy = jest.spyOn(console, 'info').mockImplementation();
-
-      // global Date override to handle timestamp generation
-      // stolen from here: https://github.com/facebook/jest/issues/2234#issuecomment-294873406
-      const DATE_TO_USE = new Date('2022');
-      // @ts-expect-error we're just overriding the constructor for tests,
-      // no need to construct everything
-      global.Date = jest.fn(() => DATE_TO_USE);
-
-      git.checkIsRepo = jest.fn(() => {
-        return Promise.resolve(true) as unknown as Response<boolean>;
-      });
-
-      git.remote = getGitRemoteMock();
-
-      fs.writeFileSync = jest.fn((f, d) => {
+      before((f, d) => {
         yamlOutput = d;
         return true;
       });
-
-      const spy = jest.spyOn(createGHAObject, 'getPkgVersion');
-      spy.mockReturnValue('7.8.9');
-
-      process.env.TEST_CREATEGHA = 'true';
     });
 
     afterEach(() => {
-      configstore.clear();
-      delete process.env.TEST_CREATEGHA;
-      jest.clearAllMocks();
+      after();
     });
 
     it('should create GHA workflow (create spec)', async () => {
