+++ __tests__/lib/createGHA.test.ts
@@ -23,7 +23,7 @@ import createGHA, {
   git,
   rdmeVersionMajor,
 } from '../../src/lib/createGHA';
-import * as createGHAObject from '../../src/lib/createGHA';
+import { after, before } from '../helpers/get-gha-setup';
 import getGitRemoteMock from '../helpers/get-git-mock';
 import ghaWorkflowSchema from '../helpers/github-workflow-schema.json';
 import '../helpers/jest.matchers';
@@ -41,35 +41,16 @@ describe('#createGHA', () => {
   beforeEach(() => {
     consoleInfoSpy = jest.spyOn(console, 'info').mockImplementation();
 
-    // global Date override to handle timestamp generation
-    // stolen from here: https://github.com/facebook/jest/issues/2234#issuecomment-294873406
-    const DATE_TO_USE = new Date('2022');
-    // @ts-expect-error we're just overriding the constructor for tests,
-    // no need to construct everything
-    global.Date = jest.fn(() => DATE_TO_USE);
-
-    git.checkIsRepo = jest.fn(() => {
-      return Promise.resolve(true) as unknown as Response<boolean>;
-    });
-
-    git.remote = getGitRemoteMock();
-
-    fs.writeFileSync = jest.fn((f, d) => {
+    before((f, d) => {
       yamlOutput = d;
       return true;
     });
-
-    const spy = jest.spyOn(createGHAObject, 'getPkgVersion');
-    spy.mockReturnValue('7.8.9');
-
-    process.env.TEST_CREATEGHA = 'true';
   });
 
   afterEach(() => {
-    configstore.clear();
+    after();
+
     consoleInfoSpy.mockRestore();
-    delete process.env.TEST_CREATEGHA;
-    jest.clearAllMocks();
     process.chdir(testWorkingDir);
   });
